api/requirements.txt
Django>=3.2.4,<3.3
djangorestframework>=3.12.4,<3.13
psycopg2>=2.8.6,<2.9
djangorestframework-simplejwt
drf-spectacular>=0.15.1,<0.16
django-cors-headers>=3.7.0,<3.8
pandas>=1.2.4,<1.3
openpyxl>=3.0.7,<3.1
numpy>=1.20.3,<1.21